+++ server/curatorship/src/main/kotlin/br/usp/inovacao/hubusp/curatorship/sheets/PDI.kt
@@ -1,8 +1,6 @@
 package br.usp.inovacao.hubusp.curatorship.sheets
 
-import org.valiktor.Constraint
 import org.valiktor.ConstraintViolationException
-import org.valiktor.Validator
 import org.valiktor.functions.*
 import org.valiktor.i18n.mapToMessage
 import org.valiktor.validate
@@ -69,36 +67,3 @@ data class PDI(
         }
     }
 }
-
-class PDICategory: Constraint
-
-fun <E> Validator<E>.Property<String?>.isPDICategory() = this.validate(PDICategory()) {
-    it == null || PDI.categories
-        .contains(it)
-}
-
-class Campi : Constraint
-
-fun <E> Validator<E>.Property<String?>.isCampus() = this.validate(Campi()) {
-    (it == null) || Campus
-        .all()
-        .contains(it)
-}
-
-class Unity : Constraint
-
-fun <E> Validator<E>.Property<String?>.isUnity() = this.validate(Unity()) {
-    it == null || it in Campus.allUnities()
-}
-
-class Phone : Constraint
-
-fun <E> Validator<E>.Property<String?>.isPhone() = this.validate(Phone()) {
-    (it == null) || it
-        .replace("""\D""".toRegex(), "")
-        .matches("""^\d{8,13}$""".toRegex())
-}
-
-class ValidationException(
-    val messages: Iterable<String>
-) : RuntimeException(messages.joinToString("|"))
