+++ server/curatorship/src/main/kotlin/br/usp/inovacao/hubusp/curatorship/sheets/ValidationException.kt
@@ -1,0 +1,5 @@
+package br.usp.inovacao.hubusp.curatorship.sheets
+
+class ValidationException(
+    val messages: Iterable<String>
+) : RuntimeException(messages.joinToString("|"))
