+++ server/curatorship/src/main/kotlin/br/usp/inovacao/hubusp/curatorship/sheets/constraints.kt
@@ -1,0 +1,33 @@
+package br.usp.inovacao.hubusp.curatorship.sheets
+
+import org.valiktor.Constraint
+import org.valiktor.Validator
+
+class PDICategory: Constraint
+
+fun <E> Validator<E>.Property<String?>.isPDICategory() = this.validate(PDICategory()) {
+    it == null || PDI.categories
+        .contains(it)
+}
+
+class Phone : Constraint
+
+fun <E> Validator<E>.Property<String?>.isPhone() = this.validate(Phone()) {
+    (it == null) || it
+        .replace("""\D""".toRegex(), "")
+        .matches("""^\d{8,13}$""".toRegex())
+}
+
+class Campi : Constraint
+
+fun <E> Validator<E>.Property<String?>.isCampus() = this.validate(Campi()) {
+    (it == null) || Campus
+        .all()
+        .contains(it)
+}
+
+class Unity : Constraint
+
+fun <E> Validator<E>.Property<String?>.isUnity() = this.validate(Unity()) {
+    it == null || it in Campus.allUnities()
+}
